Harden VLM evaluation integration
Expose the reusable evaluator from a non-CLI module so package imports follow the public API convention without preloading the documented module entry point. Normalize optional post-MIP mappings before profile registration.
